feat: ajustes gerais no projeto - Removido botão e código relacionado ao chat de suporte - Modificações diversas no CSS para ajustes visuais - Renomeados botões para melhor clareza ao usuário - Removido MongoDB, pois não haverá mais monitoramento de validade da chave

+++ Renderer/services/renderer_relogio.js
@@ -21,7 +21,7 @@ function atualizarRelogio() {
     relogioElement.appendChild(spanSeg);
 
     // Formatar e atualizar data
-    const opcoesData = { weekday: 'long', day: 'numeric', month: 'short' };
+    const opcoesData = { weekday: 'long', day: 'numeric', month: 'long' };
     const dataFormatada = dataAtual.toLocaleDateString('pt-BR', opcoesData)
         .replace('.', ''); // Remover o ponto do mês
